Personal profile

+++ Course station/Course station/Controllers/PersonalProfileController.cs
@@ -1,10 +1,12 @@
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
 using Course_station.Models;
+using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.EntityFrameworkCore;
 
 namespace Course_station.Controllers
 {
+   
     public class PersonalProfileController : Controller
     {
         private readonly ApplicationDbContext _context;
@@ -14,31 +16,17 @@ namespace Course_station.Controllers
             _context = context;
         }
 
-        // GET: PersonalProfile
-        public async Task<IActionResult> Index()
+        // GET: Learners/{learnerId}/PersonalProfile
+        [HttpGet]
+        public async Task<IActionResult> Index(int learnerId)
         {
-            return View(await _context.PersonalProfiles.ToListAsync());
+            var personalProfiles = await _context.PersonalProfiles
+                .Where(p => p.LearnerId == learnerId)
+                .ToListAsync();
+            ViewBag.LearnerId = learnerId;
+            return View(personalProfiles);
         }
 
-        // GET: PersonalProfile/Details/5
-        public async Task<IActionResult> Details(int? id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var personalProfile = await _context.PersonalProfiles
-                .FirstOrDefaultAsync(m => m.ProfileId == id);
-            if (personalProfile == null)
-            {
-                return NotFound();
-            }
-
-            return View(personalProfile);
-        }
-
-        // GET: PersonalProfile/Create/{learnerId}
         [HttpGet("PersonalProfile/Create/{learnerId}")]
         public IActionResult Create(int learnerId)
         {
@@ -49,24 +37,28 @@ namespace Course_station.Controllers
             return View(personalProfile);
         }
 
-        // POST: PersonalProfile/Create
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("PreferedContentType,EmotionalState,PersonalityType,LearnerId")] PersonalProfile personalProfile)
         {
+            if (personalProfile.LearnerId == 0)
+            {
+                ModelState.AddModelError("LearnerId", "LearnerId is required.");
+            }
+
             if (ModelState.IsValid)
             {
                 _context.Add(personalProfile);
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
             }
+            // Console.WriteLine(personalProfile.GetType()); // Should output PersonalProfile
             return View(personalProfile);
         }
 
-       
-
-        // GET: PersonalProfile/Delete/5
-        public async Task<IActionResult> Delete(int? id)
+        // GET: Learners/{learnerId}/PersonalProfile/Delete/5
+        [HttpGet("Delete/{id}")]
+        public async Task<IActionResult> Delete(int learnerId, int? id)
         {
             if (id == null)
             {
@@ -74,29 +66,29 @@ namespace Course_station.Controllers
             }
 
             var personalProfile = await _context.PersonalProfiles
-                .FirstOrDefaultAsync(m => m.ProfileId == id);
+                .FirstOrDefaultAsync(m => m.ProfileId == id && m.LearnerId == learnerId);
             if (personalProfile == null)
             {
                 return NotFound();
             }
 
+            ViewBag.LearnerId = learnerId;
             return View(personalProfile);
         }
 
-        // POST: PersonalProfile/Delete/5
-        [HttpPost, ActionName("Delete")]
+        // POST: Learners/{learnerId}/PersonalProfile/Delete/5
+        [HttpPost("Delete/{id}"), ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(int id)
+        public async Task<IActionResult> DeleteConfirmed(int learnerId, int id)
         {
-            var personalProfile = await _context.PersonalProfiles.FindAsync(id);
-            _context.PersonalProfiles.Remove(personalProfile);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
-        }
-
-        private bool PersonalProfileExists(int id)
-        {
-            return _context.PersonalProfiles.Any(e => e.ProfileId == id);
+            var personalProfile = await _context.PersonalProfiles
+                .FirstOrDefaultAsync(m => m.ProfileId == id && m.LearnerId == learnerId);
+            if (personalProfile != null)
+            {
+                _context.PersonalProfiles.Remove(personalProfile);
+                await _context.SaveChangesAsync();
+            }
+            return RedirectToAction(nameof(Index), new { learnerId });
         }
     }
 }
